LayoutTests/fast/viewport/viewport-26.html

<head>
    <title>Viewport META with "minimum-scale=8, maximum-scale=9".</title>
    <meta name="viewport" content="minimum-scale=8, maximum-scale=9">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                alert(internals.configurationForViewport(document, 1, 320, 480, 320, 352));
            }
        }
    </script>
</head>
<body onload="test();">
