LayoutTests/fast/viewport/viewport-26-expected.txt

ALERT: viewport size 980x1078 scale 8 with limits [8, 9] and userScalable true
